requirements.txt
aniso8601==9.0.1
attrs==22.2.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cachetools==4.2.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2023.3
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.16.0
SQLAlchemy==1.4.0
twilio==6.38.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.0.3
